Update README.md
There is a monkey sitting some distance away from you, and your task is to throw a banana at it. You will be asked to input an initial angle and velocity. Then, the throwing curve will be calculated by approximating the linear ODE using RK2. You have infinite chances to hit the target in case you miss, with an option to get help if you get stuck. All parameters can be edited for customization.